Problem: CQEngine has optimizations based on type checking
At certain points, CQEngine will attempt to do something like `instanceof
SimpleAttribute` or `instanceof StandingQueryAttribute`
to perform certain optimizations. Our hierarchy does not fit into
CQEngine's and therefore we're missing all these optimizations, and
this will only get worse when other attribute types will be implemented
in ES4J.

Solution: refit our attribtues into CQEngine's hierarchy.

There was another attempt to solve it through extraction of attribute
interfaces (https://github.com/npgall/cqengine/pull/82), but apparently
it will break things for other users and introduce more confusion.

The solution in this patch is not perfect as it requires careful following
of the layout and copying of necessary steps in each attribute (something
that was previously done by AbstractAttribute), but it does solve the problem.
